requirements.txt.txt
sklearn-surprise
streamlit
pandas
numpy